chore: rename back to jolt
