+++ zkml-jolt-core/src/jolt/mod.rs
@@ -1,23 +1,22 @@
-//! A state-of-the-art zkVM, called Jolt, which turns almost everything a VM does into reads and writes to memory.
-//! This includes the “fetch-decode-execute” logic of the VM.
-
 pub mod bytecode;
+pub mod execution_trace;
 pub mod instruction;
 pub mod instruction_lookups;
-pub mod lookup_trace;
 pub mod r1cs;
-pub mod registers;
-pub mod witness;
+pub mod sparse_dense_shout;
+pub mod tensor_heap;
 
 use crate::jolt::{
     bytecode::{BytecodePreprocessing, BytecodeProof},
+    execution_trace::JoltONNXCycle,
     instruction_lookups::LookupsProof,
     r1cs::{
         constraints::{JoltONNXConstraints, R1CSConstraints},
         spartan::UniformSpartanProof,
     },
-    registers::RegistersTwistProof,
+    tensor_heap::TensorHeapTwistProof,
 };
+use execution_trace::WORD_SIZE;
 use jolt_core::{
     field::JoltField,
     poly::{
@@ -26,8 +25,7 @@ use jolt_core::{
     },
     utils::{errors::ProofVerifyError, transcript::Transcript},
 };
-use lookup_trace::WORD_SIZE;
-use onnx_tracer::trace_types::{ONNXCycle, ONNXInstr};
+use onnx_tracer::{constants::MAX_TENSOR_SIZE, trace_types::ONNXInstr};
 use serde::{Deserialize, Serialize};
 use std::marker::PhantomData;
 
@@ -83,7 +81,7 @@ where
     pub trace_length: usize,
     bytecode: BytecodeProof<F, ProofTranscript>,
     instruction_lookups: LookupsProof<WORD_SIZE, F, PCS, ProofTranscript>,
-    registers: RegistersTwistProof<F, ProofTranscript>,
+    tensor_heap: TensorHeapTwistProof<F, ProofTranscript>,
     r1cs: UniformSpartanProof<F, ProofTranscript>,
     _p: PhantomData<PCS>,
 }
@@ -119,19 +117,26 @@ where
     #[tracing::instrument(skip_all, name = "Jolt::prove")]
     pub fn prove(
         mut preprocessing: JoltProverPreprocessing<F, PCS, ProofTranscript>,
-        mut trace: Vec<ONNXCycle>,
+        mut trace: Vec<JoltONNXCycle>,
     ) -> Self {
         let trace_length = trace.len();
         println!("Trace length: {trace_length}");
         F::initialize_lookup_tables(std::mem::take(&mut preprocessing.field));
         // pad trace to the next power of two
-        trace.resize(trace.len().next_power_of_two(), ONNXCycle::no_op());
+        trace.resize(trace.len().next_power_of_two(), JoltONNXCycle::no_op());
         let padded_trace_length = trace.len();
-        let ram_addresses: Vec<usize> = trace.iter().map(|cycle| cycle.td() + 1).collect();
-        let ram_K = ram_addresses.iter().max().unwrap().next_power_of_two();
+        let tensor_heap_addresses: Vec<usize> = trace
+            .iter()
+            .map(|cycle| cycle.td_write().0.last().unwrap() + 1)
+            .collect();
+        let tensor_heap_K = tensor_heap_addresses
+            .iter()
+            .max()
+            .unwrap()
+            .next_power_of_two();
         let K = [
             preprocessing.shared.bytecode.code_size,
-            ram_K,
+            tensor_heap_K,
             1 << 16, // K for instruction lookups Shout
         ]
         .into_iter()
@@ -164,19 +169,19 @@ where
                 &mut opening_accumulator,
                 &mut transcript,
             );
-        let registers_snark = RegistersTwistProof::prove(
+        let bytecode_snark =
+            BytecodeProof::prove(&preprocessing.shared.bytecode, &trace, &mut transcript);
+        let tensor_heap_snark = TensorHeapTwistProof::prove(
             &preprocessing,
             &trace,
-            ram_K,
+            tensor_heap_K,
             &mut opening_accumulator,
             &mut transcript,
         );
-        let bytecode_snark =
-            BytecodeProof::prove(&preprocessing.shared.bytecode, &trace, &mut transcript);
         JoltSNARK {
             trace_length,
             r1cs: r1cs_snark,
-            registers: registers_snark,
+            tensor_heap: tensor_heap_snark,
             instruction_lookups: instruction_lookups_snark,
             bytecode: bytecode_snark,
             _p: PhantomData,
@@ -202,14 +207,14 @@ where
             .map_err(|e| ProofVerifyError::SpartanError(e.to_string()))?;
         self.instruction_lookups
             .verify(&mut opening_accumulator, &mut transcript)?;
-        self.registers.verify(
-            padded_trace_length,
-            &mut opening_accumulator,
-            &mut transcript,
-        )?;
         self.bytecode.verify(
             &preprocessing.shared.bytecode,
-            self.trace_length,
+            padded_trace_length,
+            &mut transcript,
+        )?;
+        self.tensor_heap.verify(
+            padded_trace_length * MAX_TENSOR_SIZE,
+            &mut opening_accumulator,
             &mut transcript,
         )?;
         Ok(())
@@ -217,15 +222,15 @@ where
 }
 
 #[cfg(test)]
-mod e2e_tests {
-
+mod tests {
     use crate::{
-        jolt::{JoltProverPreprocessing, JoltSNARK},
+        jolt::{JoltProverPreprocessing, JoltSNARK, execution_trace::jolt_execution_trace},
         program::ONNXProgram,
     };
     use ark_bn254::Fr;
-    use jolt_core::poly::commitment::dory::DoryCommitmentScheme;
-    use jolt_core::utils::transcript::KeccakTranscript;
+    use jolt_core::{
+        poly::commitment::dory::DoryCommitmentScheme, utils::transcript::KeccakTranscript,
+    };
     use log::info;
     use onnx_tracer::{
         custom_addsubmul_model, logger::init_logger, model, scalar_addsubmul_model, tensor::Tensor,
@@ -234,6 +239,27 @@ mod e2e_tests {
     use std::{collections::HashMap, fs::File, io::Read};
 
     type PCS = DoryCommitmentScheme<KeccakTranscript>;
+
+    #[test]
+    fn test_custom_addsubmul() {
+        // --- Preprocessing ---
+        let custom_addsubmul_model = custom_addsubmul_model();
+        let program_bytecode = onnx_tracer::decode_model(custom_addsubmul_model.clone());
+        let pp: JoltProverPreprocessing<Fr, PCS, KeccakTranscript> =
+            JoltSNARK::prover_preprocess(program_bytecode);
+
+        // --- Proving ---
+        // Get execution trace
+        let input = Tensor::new(Some(&[10, 20, 30, 40]), &[1, 4]).unwrap();
+        let raw_trace = onnx_tracer::execution_trace(custom_addsubmul_model, &input);
+        let execution_trace = jolt_execution_trace(raw_trace);
+        println!("Execution trace: {execution_trace:#?}");
+        let snark: JoltSNARK<Fr, PCS, KeccakTranscript> =
+            JoltSNARK::prove(pp.clone(), execution_trace);
+
+        // --- Verification ---
+        snark.verify((&pp).into()).unwrap();
+    }
 
     #[test]
     fn test_scalar_addsubmul() {
@@ -246,32 +272,15 @@ mod e2e_tests {
 
         // --- Proving ---
         let input = Tensor::new(Some(&[60]), &[1]).unwrap();
-        let execution_trace = onnx_tracer::execution_trace(scalar_addsubmul_model, &input);
-        println!("Execution trace: {execution_trace:#?}");
+        let raw_trace = onnx_tracer::execution_trace(scalar_addsubmul_model, &input);
+        println!("Execution trace: {raw_trace:#?}");
+        let execution_trace = jolt_execution_trace(raw_trace);
+
         let snark: JoltSNARK<Fr, PCS, KeccakTranscript> =
             JoltSNARK::prove(pp.clone(), execution_trace);
 
         // --- Verification ---
         snark.verify((&pp).into()).unwrap();
-    }
-
-    #[test]
-    fn test_custom_addsubmul() {
-        // --- Preprocessing ---
-        let custom_addsubmul_model = custom_addsubmul_model();
-        let program_bytecode = onnx_tracer::decode_model(custom_addsubmul_model.clone());
-        let pp: JoltProverPreprocessing<Fr, PCS, KeccakTranscript> =
-            JoltSNARK::prover_preprocess(program_bytecode);
-
-        // --- Proving ---
-        let input = Tensor::new(Some(&[10, 20, 30, 40]), &[1, 4]).unwrap();
-        let execution_trace = onnx_tracer::execution_trace(custom_addsubmul_model, &input);
-        println!("Execution trace: {execution_trace:#?}");
-        // let snark: JoltSNARK<Fr, PCS, KeccakTranscript> =
-        //     JoltSNARK::prove(pp.clone(), execution_trace);
-
-        // // --- Verification ---
-        // snark.verify((&pp).into()).unwrap();
     }
 
     /// Load vocab.json into HashMap<String, (usize, i32)>
