+++ zkml-jolt-core/src/jolt/r1cs/builder.rs
@@ -1,4 +1,4 @@
-use crate::jolt::r1cs::inputs::JoltONNXR1CSInputs;
+use crate::jolt::execution_trace::{JoltONNXR1CSInputs, WitnessGenerator};
 use crate::jolt::r1cs::key::UniformR1CS;
 use jolt_core::poly::multilinear_polynomial::MultilinearPolynomial;
 use jolt_core::r1cs::builder::Constraint;
@@ -295,7 +295,7 @@ impl R1CSBuilder {
             a: a_sparse,
             b: b_sparse,
             c: c_sparse,
-            num_vars: JoltONNXR1CSInputs::num_inputs(),
+            num_vars: JoltONNXR1CSInputs::len(),
             num_rows: self.constraints.len(),
         }
     }
