+++ zkml-jolt-core/src/lib.rs
@@ -1,7 +1,6 @@
 #![allow(clippy::upper_case_acronyms)]
 #![allow(non_snake_case)]
 
-// pub mod jolt;
+pub mod jolt;
 pub mod program;
 pub mod subprotocols;
-pub mod tensor_jolt;
