--- zkml-jolt-core/src/tensor_jolt/r1cs/mod.rs
@@ -1,6 +1,0 @@
-pub mod builder;
-pub mod constraints;
-pub mod inputs;
-pub mod key;
-pub mod spartan;
-pub mod spartan_interleaved_poly;
